requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
Authlib==0.15.4
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.2.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.18
elasticsearch==7.15.1
entrypoints==0.3
fake-useragent==0.1.11
feedparser==6.0.8
Flask==2.0.2
flask-mail==0.1
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
flask-restx==0.5.1
forex-python==1.6
fuzzywuzzy==0.18.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.0
keyring==18.0.1
lamson==1.3.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
MarkupSafe==2.0.1
mock==4.0.3
multitasking==0.0.9
mysql-connector-python==8.0.27
mysqlclient==2.0.3
netifaces==0.10.4
newsapi==0.1.1
newsapi-python==0.2.6
nose==1.3.7
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.3.3
parse==1.19.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pyee==8.2.2
PyGObject==3.36.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyppeteer==0.2.6
pyquery==1.4.3
pyrsistent==0.18.0
python-apt==2.0.0+ubuntu0.20.4.5
python-daemon==2.3.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.1
python-http-client==3.3.3
python-Levenshtein==0.12.2
python-modargs==1.7
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
sendgrid==6.8.3
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
starkbank-ecdsa==1.1.1
systemd-python==234
tqdm==4.62.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
w3lib==1.22.0
wadllib==1.3.3
websockets==9.1
Werkzeug==2.0.2
xkit==0.0.0
yahoo-fin==0.8.9.1
yfinance==0.1.63
zipp==3.6.0
